requirements.txt
bs4
requests
lxml
python-decouple